Spatial relationship descriptor and lookup tables

These data.frames are used to lookup and explain which spatial relation types go with different spatial queries.

Usage

sp_rel_valid

sp_rel_lookup

Format

sp_rel_valid is a data.frame with 105 rows and 3 variables as follows:

feature_class

A feature class to be queried

query_feature_class

The feature class used to do a spatial query

sp_rel

The spatial relationships that are valid for the feature class and query_feature_class combination

sp_rel_lookup is a data.frame with 9 rows and 2 variables as follows:

sp_rel

The spatial relationship being described

description

A description of the sp_rel

Source

sp_rel_valid–Independent tests done specifically by and for arcpullr

sp_rel_lookup – https://help.arcgis.com/en/webapi/wpf/apiref/ESRI.ArcGIS.Client~ESRI.ArcGIS.Client.Tasks.SpatialRelationship.html

Details

sp_rel_lookup explains the various different types of spatial relationships available through ArcGIS REST APIs. sp_rel_valid shows which spatial relationships are valid with different geometry types being queried and used to do spatial queries